fix endpoints and styles

+++ Server/src/controllers/products.controllers.js
@@ -1,5 +1,6 @@
 import { json } from 'express';
 import connection from '../config/db.js';
+import cloudinary from "../config/cloudinary.js";
 
 
 
@@ -47,28 +48,40 @@ export const getOneproduct = async (req, res) => {
 export const createProduct = async (req, res) => {
 
     try {
-        const { name, price, description, image_url, stock } = req.body
 
-        if (!name || !price || !description || !image_url || !stock) {
+        const resultCloudinary = await cloudinary.uploader.upload(req.file.path, {
+            folder: 'e-commerce'
+        })
+
+        const { name, price, description, stock } = req.body
+
+
+
+        if (!name || !price || !description || !resultCloudinary || !stock) {
             return res.status(400).json({ message: 'All field Must be filled' })
         }
 
-        const sql = 'INSERT INTO Products (name, price, description, image_url, stock) VALUES(?,?,?,?,?)'
+        const sql = 'INSERT INTO Products (name, price, description, image_url, public_id, stock) VALUES(?,?,?,?,?,?)'
 
-        const values = [name, price, description, image_url, stock]
+        const values = [name, price, description, resultCloudinary.secure_url, resultCloudinary.public_id, stock]
 
         const [result] = await connection.execute(sql, values)
 
         res.status(201).json({
             message: 'Product was created successly',
-            productId: result.insertId,
             product: {
-                name, price, description, image_url, stock
+                id: result.insertId,
+                name,
+                price,
+                description,
+                stock,
+                image_url: resultCloudinary.secure_url,
+                public_id: resultCloudinary.public_id
             }
         })
 
     } catch (error) {
-        res.status(500), json({ message: 'Internal server error:', error })
+        res.status(500).json({ message: 'Internal server error:', error })
     }
 }
 
@@ -136,21 +149,32 @@ export const deleteProducts = async (req, res) => {
 
     try {
 
-        const { id } = req.params
+        const { id } = req.params;
 
-        const sql = 'DELETE FROM Products WHERE id = ?'
 
-        const value = []
-        value.push(id)
+        const [rows] = await connection.execute(
+            "SELECT public_id FROM Products WHERE id = ?",
+            [id]
+        );
 
-        const [result] = await connection.execute(sql, value)
-
-        if (result.affectedRows === 0) {
-            return res.status(404).json({ message: 'Products not faund' })
+        if (rows.length === 0) {
+            return res.status(404).json({ message: "Product not found" });
         }
 
-        res.status(201).json({ message: 'Product was deleted successly', })
+        const publicId = rows[0].public_id;
 
+
+        if (publicId) {
+            await cloudinary.uploader.destroy(publicId);
+        }
+
+
+        const [result] = await connection.execute(
+            "DELETE FROM Products WHERE id = ?",
+            [id]
+        );
+
+        res.status(200).json({ message: "Product was deleted successfully" });
 
     } catch (error) {
         res.status(500).json({ message: 'Internar server error:', error })
